refactor: Separate Color and Type enum
Separated Color and Type enum into independent files.
