+++ src/main/java/pieces/Color.java
@@ -1,0 +1,7 @@
+package pieces;
+
+public enum Color {
+    WHITE,
+    BLACK,
+    NO_COLOR
+}
